added coredata entities and changed int64 to int16 and doubles to floats

+++ CalculatedPoints/CalculatingLoadingManager.swift
@@ -14,7 +14,7 @@ import Foundation
 ///
 /// The class has a private cancellables property, which is a set of AnyCancellable instances that manage the lifetimes of Combine framework's publishers. The class also has a private progressSubject property, which is a PassthroughSubject instance that can be used to send progress values.
 ///
-/// The updateProgress function is used to update the progress property. It takes a Double value as a parameter, which represents the progress value. When called, the function sends the value using the progressSubject. The PassthroughSubject forwards the value to the progress property, which is updated.
+/// The updateProgress function is used to update the progress property. It takes a Float value as a parameter, which represents the progress value. When called, the function sends the value using the progressSubject. The PassthroughSubject forwards the value to the progress property, which is updated.
 class CalculatingLoadingManager: ObservableObject {
     // Create a shared instance of the class (singleton pattern) to be used throughout the application
     static let shared = CalculatingLoadingManager()
@@ -22,16 +22,16 @@ class CalculatingLoadingManager: ObservableObject {
     // A set to store the Combine framework's cancellable instances to manage their lifetimes
     private var cancellables = Set<AnyCancellable>()
 
-    // Define a @Published property called progress of type Double, which is initialized to 0.0
+    // Define a @Published property called progress of type Float, which is initialized to 0.0
     // When this property is updated, SwiftUI will automatically update the UI
-    @Published var progress: Double = 0.0
+    @Published var progress: Float = 0.0
 
     @Published var nameLoading: String = ""
 
-    // Create a PassthroughSubject instance of type Double and Never, which can be used to send progress values
-    private let progressSubject = PassthroughSubject<Double, Never>()
+    // Create a PassthroughSubject instance of type Float and Never, which can be used to send progress values
+    private let progressSubject = PassthroughSubject<Float, Never>()
 
-    // Create a PassthroughSubject instance of type Double and Never, which can be used to send progress values
+    // Create a PassthroughSubject instance of type Float and Never, which can be used to send progress values
     private let nameLoadingSubject = PassthroughSubject<String, Never>()
 
     // The private initializer for the class
@@ -48,8 +48,8 @@ class CalculatingLoadingManager: ObservableObject {
             .assign(to: &$nameLoading)
     }
 
-    // Define a function called updateProgress that takes a Double value as a parameter
-    func updateProgress(_ value: Double) {
+    // Define a function called updateProgress that takes a Float value as a parameter
+    func updateProgress(_ value: Float) {
         // Send the value using the progressSubject, which will update the progress property and the UI
         progressSubject.send(value > 1 ? 1 : value)
     }
